% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemunits.R
\name{as_chemunits}
\alias{as_chemunits}
\title{convert to chemunits}
\usage{
as_chemunits(
  x = double(),
  ...,
  convert_to_preferred = FALSE,
  auto_scale = TRUE
)
}
\arguments{
\item{...}{passed on to \code{\link[units:units]{units::as_units()}}}
}
\description{
convert to chemunits
}
\keyword{internal}
