% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_scaling.R
\name{auto_scale_units}
\alias{auto_scale_units}
\title{automatically scale numerators to the best udunits prefix if x has units
that are listed in the scalable_units}
\usage{
auto_scale_units(x, scalable_units = get_chemunits_option("auto_scale_units"))
}
\arguments{
\item{x}{a units object}

\item{scalable_units}{a vector of scalable units}
}
\description{
automatically scale numerators to the best udunits prefix if x has units
that are listed in the scalable_units
}
