% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemunits-package.R
\docType{package}
\name{chemunits-package}
\alias{chemunits}
\alias{chemunits-package}
\title{chemunits: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

The chemunits package augments the units package for use in chemistry calculations.

Resources:
\itemize{
\item Website for the chemunits package: \url{https://chemunits.kopflab.org}
\item Create a chemunit: \code{\link[=set_cu]{set_cu()}}
\item Package options: \link{chemunits_options}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kopflab/chemunits}
  \item \url{http://www.kopflab.org/chemunits/}
  \item Report bugs at \url{https://github.com/kopflab/chemunits/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Kopf \email{sebastian.kopf@colorado.edu} (\href{https://orcid.org/0000-0002-2044-0201}{ORCID}) [copyright holder]

}
