% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options-package.R
\name{chemunits_options}
\alias{chemunits_options}
\alias{get_chemunits_option}
\alias{get_pkg_options}
\title{Package options}
\usage{
chemunits_options(...)

get_chemunits_option(x)
}
\arguments{
\item{...}{set package options, syntax identical to \code{\link[=options]{options()}}}

\item{x}{name of the option to retrieve}
}
\description{
These options are best set via \code{\link[=chemunits_options]{chemunits_options()}} and queried via \code{\link[=get_chemunits_option]{get_chemunits_option()}}. However, the base functions \code{\link[=options]{options()}} and \code{\link[=getOption]{getOption()}} work as well but require a \code{chemunits.} prefix (the package name and a dot) for the option name. Setting an option to a value of \code{NULL} means that the default is used.
}
\section{Functions}{
\itemize{
\item \code{chemunits_options()}: set option values

\item \code{get_chemunits_option()}: retrieve the current value of an option

}}
\section{Options for the chemunits package}{

\itemize{
\item \code{default_units} (character vector of units): the default units that new units (i.e. resulting from calculations) should be converted to. Default: \code{"1"} (i.e. units that cancel themselves and should be dimensionless are converted to unitless). The order does not matter except if units can be interconverted from each other (e.g. \code{"L"} and \code{"m^3"}) in which case the first one will be used (with a warning).
\item \code{auto_scale_units} (character vector of units): the units whose best SI prefix (e.g. nano/n, micro/u, milli/m, kilo/k, etc) should be determined automatically based on the median value of a vector in this unit. Default: \code{character(0)} (i.e. no units are automatically scaled).
}
}

\examples{
# Default setting(s):
old <- chemunits_options()
old # original options
get_chemunits_option("default_units")

# with this, there is no default unit defined for energy or volume
# so the units are not simplified
set_cu(100, "W") * set_cu(1, "s")
set_cu(1, "m") * set_cu(1, "mm^2")

# however the default does simplify dimensionless units
set_cu(1, "W*s") / set_cu(1, "J")

# Change for the duration of the session:
chemunits_options(default_units = c("J", "L"))

# now W*s is recognized as energy and cm mm^2 as a volume and both are
# automatically converted to the specified default units
set_cu(100, "W") * set_cu(1, "s")
set_cu(1, "m") * set_cu(1, "mm^2")

# however since "1" was no longer included in the `default_units`, 
# dimensionless units were no longer simplified (thus it is advisable
# to always include "1" in the `default_units` option)
set_cu(1, "W*s") / set_cu(1, "J")

# Restore original values
chemunits_options(old)

}
