% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_units.R
\name{convert_to_default_units}
\alias{convert_to_default_units}
\title{automatically convert to default units if any fit (\code{defaults} would be an option in units_options)}
\usage{
convert_to_default_units(x, defaults = get_chemunits_option("default_units"))
}
\arguments{
\item{x}{units object}

\item{defaults}{default units to convert to}
}
\description{
automatically convert to default units if any fit (\code{defaults} would be an option in units_options)
}
