% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemunits.R
\name{get_chemvalue}
\alias{get_chemvalue}
\alias{get_cv}
\title{\code{chemunits} vector}
\usage{
get_chemvalue(x, units)

get_cv(x, units)
}
\arguments{
\item{x}{a chemunits vector}

\item{units}{the units}
}
\value{
A \code{double} vector
}
\description{
\code{chemunits} vector
}
\section{Functions}{
\itemize{
\item \code{get_cv()}: short-form for the \code{get_chemvalue} function

}}
