% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemunits.R
\name{get_units}
\alias{get_units}
\title{Retrieve units}
\usage{
get_units(x, prefix = group[1], suffix = group[2], group = c("", ""))
}
\arguments{
\item{x}{a chemunits vector}

\item{prefix}{the prefix text for the units. Default: \code{""} (no prefix).}

\item{suffix}{the suffix text for the units. Default: \code{""} (no suffix).}

\item{group}{parameter to specific the prefix and suffix text as a vector of length 2. This is helpful when using .e.g the prefix/suffix from the \code{\link[units:units_options]{units::units_options()}} \code{group} for this: \code{group = units_options("group")}.}
}
\value{
a text representation of the units
}
\description{
This function retrieves a text representation of the units of a units/chemunits object. To get the symbolic units, use \code{\link[units:units]{units::units()}} instead.
}
\examples{
set_cu(1, "m/s") |> get_units()
set_cu(1, "m/s") |> get_units(prefix = "(", suffix = ")")
set_cu(1, "m/s") |> get_units(group = units_options("group"))
}
