% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemunits.R
\name{make_units_explicit}
\alias{make_units_explicit}
\title{Make units explicit}
\usage{
make_units_explicit(
  df,
  ...,
  prefix = paste0(" ", group[1]),
  suffix = group[2],
  group = units_options("group")
)
}
\arguments{
\item{df}{the data frame in which to make the units explicit}

\item{...}{which columns to make explicit ([dplyr::select())] syntax). If none provided, makes all columns explicit.

[dplyr::select())]: R:dplyr::select())}

\item{prefix}{the prefix text for the units. Default: \code{" ["} (based on \code{\link[units:units_options]{units::units_options()}}).}

\item{suffix}{the suffix text for the units. Default: \code{"]"} (based on \code{\link[units:units_options]{units::units_options()}}).}

\item{group}{parameter to specify the prefix and suffix text as a vector of length 2. This is helpful because it allows compatiblity with the \code{\link[units:units_options]{units::units_options()}} \code{group} option.}
}
\value{
a text representation of the units
}
\description{
This function is intended for data frames / tibbles only and makes the specified columns explicit with the target units in the column name.
}
\examples{
# data frame with quantities
df <- tibble(
  weight = set_cu(1:5, "mg"),
  vol = set_cu(20, "mL"),
  mw = set_cu(500, "g/mol"),
  amount = weight / mw,
  conc = amount / vol
)

# make all units columns explicit
df |> make_units_explicit()

# change the prefix and suffix
df |> make_units_explicit(prefix = ".", suffix = "")

# make a specific subset of columns explicit (using tidy select syntax)
df |> make_units_explicit(weight:mw) # from weight to mw
df |> make_units_explicit(-mw) # all but mw
}
\concept{quantity functions}
