% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemunits.R
\name{set_chemunits}
\alias{set_chemunits}
\alias{set_cu}
\title{\code{chemunits} vector}
\usage{
set_chemunits(x, units, convert_to_preferred = FALSE, auto_scale = TRUE)

set_cu(x, units, convert_to_preferred = FALSE, auto_scale = TRUE)
}
\arguments{
\item{x}{A numeric vector}

\item{units}{units}

\item{convert_to_preferred}{whether to convert to default (if there is default)}

\item{auto_scale}{whether to auto-scale (if units are scalable)}
}
\value{
An S3 vector of class \code{chemunits}.
}
\description{
\code{chemunits} vector
}
\section{Functions}{
\itemize{
\item \code{set_cu()}: short-form for the \code{set_chemunits} function

}}
\examples{
set_chemunits(c(0.25, 0.5, 0.75))
}
